requirements.txt
asgiref==3.4.1
Django==3.2.8
django-extensions==3.1.3
djangorestframework==3.12.4
drf-generators==0.5.0
Owlready2==0.34
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
